% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allan.R
\name{allan_factor}
\alias{allan_factor}
\title{Compute Allan Factor}
\usage{
allan_factor(fin, numint)
}
\arguments{
\item{fin}{Binary event vector}

\item{numint}{Number of intervals}
}
\value{
List with allan value and af vector
}
\description{
Compute Allan Factor
}
