% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{biobehavioral_interactions}
\alias{biobehavioral_interactions}
\title{Mother-Child Biobehavioral Interactions}
\format{
A data frame with 16705 rows and 18 columns:
\describe{
  \item{DyadID}{Dyad identifier}
  \item{dc_time}{Timestamp in seconds}
  \item{mom_IBI_Lego}{Continuous inter-beat intervals for mothers}
  \item{child_IBI_Lego}{Continuous inter-beat intervals for children}
  \item{mom_RSA_Lego}{Continuous respiratory sinus arrhythmia for mothers}
  \item{child_RSA_Lego}{Continuous respiratory sinus arrhythmia for children}
  \item{mom_HbO_dlpfc_l}{Oxygenated hemoglobin aggregated across the left dorsolateral prefrontal cortex for mothers}
  \item{mom_HbO_dlpfc_r}{Oxygenated hemoglobin aggregated across the right dorsolateral prefrontal cortex for mothers}
  \item{mom_HbR_dlpfc_l}{Deoxygenated hemoglobin aggregated across the left dorsolateral prefrontal cortex for mothers}
  \item{mom_HbR_dlpfc_r}{Deoxygenated hemoglobin aggregated across the right dorsolateral prefrontal cortex for mothers}
  \item{child_HbO_dlpfc_l}{Oxygenated hemoglobin aggregated across the left dorsolateral prefrontal cortex for children}
  \item{child_HbO_dlpfc_r}{Oxygenated hemoglobin aggregated across the right dorsolateral prefrontal cortex for children}
  \item{child_HbR_dlpfc_l}{Deoxygenated hemoglobin aggregated across the left dorsolateral prefrontal cortex for children}
  \item{child_HbR_dlpfc_r}{Deoxygenated hemoglobin aggregated across the right dorsolateral prefrontal cortex for children}
  \item{NA_child}{Children's negative affect; coded as 1 when present and 0 when not present}
  \item{PA_child}{Children's positive affect; coded as 1 when present and 0 when not present}
  \item{NA_Parent}{Mothers' negative affect; coded as 1 when present and 0 when not present}
  \item{PA_Parent}{Mothers' positive affect; coded as 1 when present and 0 when not present}
}
}
\source{
Steffen, G., & Abney, D. (2025, October 24). Coupling of rhythms in prefrontal cortex and autonomic nervous system in school-age children. Retrieved from osf.io/mqfbp
}
\usage{
biobehavioral_interactions
}
\description{
A sample dataset for exploring the DORA app,
containing physiology and behavior of mothers
and their 5-7 year old children across a 5-minute
stress task.
}
\keyword{datasets}
