% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allan.R
\name{compute_allan_deviation}
\alias{compute_allan_deviation}
\title{Compute Allan Deviation}
\usage{
compute_allan_deviation(data, rate = 1, type = c("frequency", "phase"))
}
\arguments{
\item{data}{Numeric vector}

\item{rate}{Sampling rate in Hz}

\item{type}{One of "frequency" or "phase"}
}
\value{
Data frame with tau and adev columns
}
\description{
Compute Allan Deviation
}
