% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dataset1_object_play}
\alias{dataset1_object_play}
\title{Infant Object Play}
\format{
A data frame with 16032 rows and 58 columns:
\describe{
  \item{id}{Participant identifier}
  \item{id_onset}{Start onset for each participant in milliseconds}
  \item{id_offset}{End offset for each participant in milliseconds}
  \item{agemo}{Participant age in months}
  \item{sex}{Participant sex, m = male; f = female}
  \item{onset}{The onset or start of every object interaction bout in milliseconds}
  \item{offset}{The offset or end of every object interaction bout in milliseconds}
  \item{bobj}{The object engagement status: 'o' = object, '.' = out of view}
  \item{objcat}{Category of object interaction (e.g., toy)}
}
}
\source{
Tamis-LeMonda, C., Adolph, K. & Herzberg, O. (2020). Infant exuberant object play at home: Immense amounts of time-distributed, variable practice. Databrary. Retrieved May 14, 2026 from https://databrary.org/volume/1118.
}
\usage{
object_play
}
\description{
A sample dataset for exploring the DORA app,
containing object interaction bouts across an hour for 50 infants.
}
\keyword{datasets}
