% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dataset4_cradling_diaries}
\alias{dataset4_cradling_diaries}
\title{Gahvora Cradling Diaries}
\format{
A data frame with 6073 rows and 105 columns:
\describe{
  \item{id}{Participant identifier}
  \item{agegrp}{The age group that the subject is in. Format should be digits only (0, 4, 8, 12, 16, 20, 24).}
  \item{sex}{Sex of participant; m = male, f = female}
  \item{marital}{The marital status of the subject’s mother. s = single, m = married, w = widowed, d = divorced, p = separated}
  \item{region}{The region where data collection occurred (e.g., k = Khatlon, r = Rasht, d = Dushanbe)}
  \item{numkids}{The number of children the mother has}
  \item{mworktyp}{Mother's work type outside the home}
  \item{fworktyp}{Father's work type outside the home}
  \item{tempout}{The temperature outdoors (in degrees Celsius)}
  \item{tempin}{The temperature indoors (in degrees Celsius)}
}
}
\source{
Karasik, L., Tamis-LeMonda, C. & Adolph, K. (2014). The Ties That Bind: Cradling in Tajikistan. Databrary. Retrieved May 14, 2026 from https://www.databrary.org/volume/11.
}
\usage{
cradling_diaries
}
\description{
A sample dataset for exploring the DORA app,
containing detailed coding of gahvora cradle use,
daily routines, including location and feeding,
and structured tasks, including behaviorally-coded
object interactions, and AIMS skills for
118 infants (aged 12, 16, 20 months).
}
\keyword{datasets}
