% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{demo_data_1}
\alias{demo_data_1}
\title{Demo time series dataset 1}
\format{
A data frame with 16032 rows and 12 columns:
\describe{
  \item{id_onset}{Start onset for each participant in milliseconds}
  \item{id_offset}{End offset for each participant in milliseconds}
  \item{id}{Participant identifier}
  \item{agomo}{Participant age in months}
  \item{sex}{Participant sex, m = male; f = female}
  \item{onset}{The onset or start of every object interaction bout in milliseconds}
  \item{onset}{The offset or end of every object interaction bout in milliseconds}
}
}
\source{
Simulated example data 1
}
\usage{
demo_data_1
}
\description{
A sample dataset for exploring the DORA app,
containing object interaction bouts across an hour for 50 infants.
}
\keyword{datasets}
