% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_dataset.R
\name{detect_dataset}
\alias{detect_dataset}
\title{Detect variable types in a dataframe}
\usage{
detect_dataset(df)
}
\arguments{
\item{df}{A data frame}
}
\value{
A list with elements: numeric, binary, time, n_rows, n_cols
}
\description{
Identifies numeric, binary, and time variables automatically.
}
