% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dora-package.R
\docType{package}
\name{dora-package}
\alias{dora}
\alias{dora-package}
\title{dora: Measure DevOps DORA Metrics}
\description{
An implementation of the key metrics for measuring fast delivery of reliable software. The DORA metrics are 1. Change Lead Time 2. Deployment Frequency 3. Mean Time to Recovery 4. Change Failure Rate
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/statik/dora}
  \item \url{https://statik.github.io/dora/}
  \item Report bugs at \url{https://github.com/statik/dora/issues}
}

}
\author{
\strong{Maintainer}: Elliot Murphy \email{elliot@elliotmurphy.com}

}
\keyword{internal}
