% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_event_windows_idx}
\alias{extract_event_windows_idx}
\title{Extract event window indices from a binary vector}
\usage{
extract_event_windows_idx(event)
}
\arguments{
\item{event}{Binary numeric vector}
}
\value{
Data frame with start and end columns
}
\description{
Extract event window indices from a binary vector
}
