% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/burstiness.R
\name{get_burstiness}
\alias{get_burstiness}
\title{Compute burstiness of a binary event vector}
\usage{
get_burstiness(vector)
}
\arguments{
\item{vector}{A binary (0/1) numeric vector}
}
\value{
Numeric burstiness value, or NA if not computable
}
\description{
Compute burstiness of a binary event vector
}
