% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_datetime_axis.R
\name{get_datetime_axis}
\alias{get_datetime_axis}
\title{Gets tick marks and plot labels for dora_app() plots}
\usage{
get_datetime_axis(time_vec, title_text, fonts)
}
\arguments{
\item{time_vec}{a vector containing a time variable}

\item{title_text}{the desired title for the time (x) axis}

\item{fonts}{a vector containing font sizes for different plot features}
}
\value{
A list of plotting labels for tickmarks, titles, and fonts
}
\description{
Calculates tick intervals that are adapted to the length of a timeseries to avoid clustered tick marks and outputs them with other formatting for the dora_app
}
