% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{music_bouts}
\alias{music_bouts}
\title{Daily Music Bouts}
\format{
A data frame with 4,798 rows and 28 columns:
\describe{
  \item{subjID}{Subject ID}
  \item{DayInSeconds}{time of day in cumulative seconds from midnight=0 of the start of the music bout}
  \item{ClocktimeSeconds}{time of day in H:M:S of the start of the music bout}
  \item{DLPspan}{was this music bout during the time span between when the DLP was first turned on and when it was last turned off ("DLP span")? 1=yes, 0=no}
  \item{DLP_On}{was this music bout from a portion of the recording when the DLPwas  turned on? 1=yes, 0=no}
  \item{DLPon_NoPrivNoOut}{was this music bout from a portion of the recording when the DLP was turned on AND not from a  privacy or outside-home section? 1=yes, 0=no}
  \item{AudSound_Edited}{was this music bout from a portion of the recording that was coded? 1=yes, 0=no}
  \item{MusicBouts}{was this music bout from a portion of the recording that was coded as a music bout? 1=yes, 0=no}
  \item{MusicBout_Smoothed}{the number of the music bout based on the ELAN coding (smoothed bouts show the numbers of the first and last ELAN bout in the set separated by "_")}
  \item{Goof}{was this music bout coded as a 'goof' bout? 1=yes, 0=no}
  \item{Live}{was this music bout coded as 'live'? 1=yes, 0=no}
  \item{Recorded}{was this music bout coded as 'recorded'? 1=yes, 0=no}
  \item{Vocal}{was this music bout coded as 'vocal'? 1=yes, 0=no}
  \item{Instrumental}{was this music bout coded as 'instrumentall'? 1=yes, 0=no}
  \item{Voices_Tunes}{Voices and tunes coded in the music bout. Tunes are linked to the voice that produced them with "_".  Multiple voice_tune combinations are separated by "QQQ"}
  \item{Voices_Unique}{the unique set of voices coded in the music bout, separated by "QQQ"}
  \item{Tunes_Unique}{the unique set of tunes coded in the music bout, separated by "QQQ"}
  \item{Tunes_Voc}{the unqiue set of tunes that were produced by a voice in the music bout, separated by "QQQ"}
  \item{Tunes_Instr}{the unique set of instrumental tunes in the music bout, separated by "QQQ"}
  \item{Goof_firstRow}{was this the first row of a goof bout? 1=yes, 0=no}
  \item{MusicBouts_noGoofs}{was this part of a music bout AND not marked as a goof? 1=yes, 0=no}
  \item{MBfirstRow}{is this the first row of a music bout? 1=yes, 0=no}
  \item{oneVoice}{does this music bout have exactly 1 unique voice? 1=yes, 0=no}
  \item{oneTune}{does this music bout have exactly 1 unique tune? 1=yes, 0=no}
  \item{LV_1V1T}{was this music bout coded as 'live', 'vocal' (but not as 'vocal' or 'instrumental'), and does it have exactly 1 unique voice and 1 unique tune? 1=yes, 0=no}
  \item{RVI_1V1T}{was this music bout coded as 'recorded', 'vocal', & 'instrumental (but not as 'live'), and does it have exactly 1 unique voice and 1 unique tune? 1=yes, 0=no}
  \item{DurationSecs}{the duration in seconds of the music bout}
  \item{PauseSecs}{the duration in seconds of the interval between the end of this music bout and the beginning of the subsequent music bout}
}
}
\source{
Mendoza, J. K., & Fausey, C. M. (2022, January 14). Everyday Music in Infancy. https://doi.org/10.17605/OSF.IO/EB9PW
}
\usage{
music_bouts
}
\description{
A sample dataset for exploring the DORA app,
containing music listening bouts from daylong recordings
of 35 infants ages 6-12 months at home. Music bouts were
fully double-coded and include 467 hours of everyday music
and its features, tunes, and voices.
}
\keyword{datasets}
