% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Launch the Time Series Explorer Shiny app}
\usage{
run_app(
  port = getOption("shiny.port"),
  host = getOption("shiny.host", "127.0.0.1")
)
}
\arguments{
\item{port}{Port to run the app on. Defaults to Shiny's default.}

\item{host}{Host address. Defaults to 127.0.0.1.}
}
\description{
Launch the Time Series Explorer Shiny app
}
