% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_id_variable}
\alias{validate_id_variable}
\title{Validate an ID variable in a dataframe}
\usage{
validate_id_variable(df, col_name)
}
\arguments{
\item{df}{A data frame}

\item{col_name}{Name of the column to validate}
}
\value{
NULL if valid, error string if not
}
\description{
Validate an ID variable in a dataframe
}
