% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allan.R
\name{compute_allan_factor_curve}
\alias{compute_allan_factor_curve}
\title{Compute Allan Factor curve}
\usage{
compute_allan_factor_curve(
  fin,
  binwidth,
  base = 2,
  powers = 10,
  start = 2,
  include_shuffled = TRUE
)
}
\arguments{
\item{fin}{Binary event vector}

\item{binwidth}{Bin width in seconds}

\item{base}{Logarithm base}

\item{powers}{Maximum power}

\item{start}{Starting power}

\item{include_shuffled}{Whether to include shuffled comparison}
}
\value{
List with actual, abcis, shuffled, slope
}
\description{
Compute Allan Factor curve
}
