% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cradling_diaries}
\alias{cradling_diaries}
\title{Demo time series dataset 4}
\format{
A data frame with X rows and Y columns:
\describe{
  \item{id}{Participant identifier}
  \item{agegrp}{he age group that the subject is in. Format should be digits only (0, 4, 8, 12, 16, 20, 24).}
  \item{sex}{Sex of participant; m = male, f = female}
  \item{marital}{The marital status of the subject’s mother. s = single, m = married, w = widowed, d, = divorced, p = separated }
}
}
\source{
Real example data 4
}
\usage{
cradling_diaries
}
\description{
A sample dataset for exploring the DORA app,
containing object interaction bouts across an hour for 50 infants.
}
\keyword{datasets}
