% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_timeseries.R
\name{expand_timeseries}
\alias{expand_timeseries}
\title{Expand interval data to continuous time series format}
\usage{
expand_timeseries(
  data,
  id_var,
  var_name,
  start_time_var,
  end_time_var,
  time_unit
)
}
\arguments{
\item{data}{A data frame containing interval data}

\item{id_var}{Name of the participant ID column}

\item{var_name}{Name of the event/activity column}

\item{start_time_var}{Name of the start time column}

\item{end_time_var}{Name of the end time column}

\item{time_unit}{Numeric time step in seconds}
}
\value{
A data frame in continuous time series format
}
\description{
Converts a dataframe with start/end time or start time and duration columns into a row-per-timepoint
format, filling gaps with zeros.
}
