% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{music_bouts}
\alias{music_bouts}
\title{Demo time series dataset 3}
\format{
A data frame with X rows and Y columns:
\describe{
  \item{subjID}{Subject ID}
  \item{DayInSeconds}{time of day in cumulative seconds from midnight=0 of the start of the music bout}
  \item{ClocktimeSeconds}{time of day in H:M:S of the start of the music bout}
  \item{DLPspan}{was this music bout during the time span between when the DLP was first turned on and when it was last turned off ("DLP span")? 1=yes, 0=no}
}
}
\source{
Real example data 3
}
\usage{
music_bouts
}
\description{
A sample dataset for exploring the DORA app,
containing music listening bouts...
}
\keyword{datasets}
